create dispatcher service and use spring RestTemplate for requests
